refactor: validations를 utils 안으로 이동
