--- src/hooks/useInput.tsx
@@ -1,10 +1,0 @@
-import { useState } from 'react';
-
-function useInput() {
-  const [value, setValue] = useState('');
-  const onChange = (e) => setValue(e.target.value);
-
-  return { value, onChange };
-}
-
-export default useInput;
